comment and change the file location
Added explanations for all code. Move the document that holds the code out of Game Assets and into the main folder.
